+++ src/lib/constants.ts
@@ -1,8 +1,8 @@
 export const CLASSNAMES = {
   /** Padding for <Main> or layout components */
-  layoutPadding: 'px-4 md:px-8 min-[100rem]:px-0',
+  layoutPadding: 'px-4 md:px-8',
   /** Force content to have full width in <Main> */
-  fullMainWidth: '-mx-4 md:-mx-8 px-4 md:px-8 min-[100rem]:px-0 min-[100rem]:mx-0',
+  fullMainWidth: '-mx-4 md:-mx-8 px-4 md:px-8',
   /** Spacing for sections */
   sectionSpacing: 'space-y-12 sm:space-y-16 lg:space-y-28',
   /** Centered container */