+++ src/lib/payload/index.ts
@@ -1,7 +1,6 @@
 export * from './can-use-dom'
 export * from './deep-merge'
 export * from './format-authors'
-export * from './format-date-time'
 export * from './generate-meta'
 export * from './get-media-url'
 export * from './get-url'